% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alpha_to_prob}
\alias{alpha_to_prob}
\title{Transform threshold into probabilities}
\usage{
alpha_to_prob(alpha, link = "logit", ...)
}
\arguments{
\item{alpha}{a vector of thresholds}

\item{link}{the link function. \code{logit} or \code{probit}}

\item{...}{other arguments passed to the appropriate cumulative distribution function (\code{pnorm} or \code{plogis})}
}
\value{
a vector of probabilities of length \code{length(alpha) + 1}
}
\description{
Transform threshold into probabilities
}
\examples{
alpha <- c(-2, -0.5, 1, 1.5)
alpha_to_prob(alpha, link = "logit")
alpha_to_prob(alpha, link = "probit")
}
