% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plotting.R
\name{cat_latent_plot}
\alias{cat_latent_plot}
\title{Shows the effect of a categorical variable on predicted ordinal probabilities}
\usage{
cat_latent_plot(
  location = 0,
  scale = 1,
  alpha = NULL,
  prob0 = NULL,
  link = "logit",
  plot = TRUE
)
}
\arguments{
\item{location}{vector of group latent locations}

\item{scale}{vector of group latent scales}

\item{alpha}{vector of threshold}

\item{prob0}{vector of baseline probabilities}

\item{link}{}

\item{plot}{}
}
\description{
The function given a vector of latent means, scales, and the thresholds or
baseline probabilities plot the latent distributions and the expected probabilities
}
