% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_ord}
\alias{dummy_ord}
\title{Create k - 1 dummy variables from an ordinal variable}
\usage{
dummy_ord(y)
}
\arguments{
\item{y}{a vector. Can be numeric, integer or (ordered) factor. If \code{y} is a factor, the function extract the underlying integer representation.}
}
\value{
a data.frame with k - 1 dummy variables.
}
\description{
Given an ordinal variable with \code{k} levels the function return a data.frame with k - 1 dummy variables.
}
\examples{
k <- 4
y <- sample(1:k, size = 20, replace = TRUE)
dummy_ord(y)
}
