% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_link}
\alias{get_link}
\title{List of probability functions}
\usage{
get_link(link = "logit")
}
\arguments{
\item{link}{the link function. \code{logit} or \code{probit}}
}
\value{
a list with the data generation function, cumulative and inverse cumulative probability function and the density function.
}
\description{
List of probability functions
}
\examples{
str(get_link("logit"))
str(get_link("probit"))
}
