% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plotting.R
\name{num_latent_plot}
\alias{num_latent_plot}
\title{Shows the effect of a numerical variable on predicted ordinal probabilities}
\usage{
num_latent_plot(
  x,
  b1,
  alpha = NULL,
  prob0 = NULL,
  at = NULL,
  link = "logit",
  size = 20,
  linewidth = 1,
  plot = TRUE
)
}
\arguments{
\item{x}{vector with the predictor}

\item{b1}{regression coefficient}

\item{alpha}{vector of thresholds}

\item{prob0}{vector of probabilities}

\item{at}{vector of \code{x} values where probabilities are calculated. If \code{NULL}, the function use \code{quantile(x, c(0.25, 0.5, 0.75))}}

\item{link}{link function (logit or probit)}

\item{size}{size of the \code{ggplot2::theme_minimal(base_size = )}}

\item{linewidth}{size of the \code{ggplot2::geom_line()}}
}
\value{
list of plots with the predicted probabilities and barplots at specific \code{x} values.
}
\description{
Given the predictor \code{x} and a single regression parameter \code{b1} the function return a plot
of the predicted probabilities P(Y = k|X). The function return also the plot of the predicted
probabilities at specific \code{x} values. By default the 25th, 50th and 75th quantiles but custom
values can be provided.
}
