% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_to_alpha}
\alias{prob_to_alpha}
\title{Transform probabilities into thresholds}
\usage{
prob_to_alpha(prob, link = "logit", ...)
}
\arguments{
\item{prob}{a vector of probabilities}

\item{link}{the link function. \code{logit} or \code{probit}}

\item{...}{other arguments passed to the appropriate inverse cumulative distribution function (\code{qnorm} or \code{qlogis})}
}
\value{
a vector of thresholds of length \code{length(prob) - 1}
}
\description{
Transform probabilities into thresholds
}
\examples{
probs <- c(0.5, 0.3, 0.1, 0.1)
prob_to_alpha(probs, link = "logit")
prob_to_alpha(probs, link = "probit")
}
