% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plotting.R
\name{show_alpha}
\alias{show_alpha}
\title{Show the thresholds and associated probabilities}
\usage{
show_alpha(prob = NULL, alpha = NULL, link = "logit", plot = TRUE)
}
\arguments{
\item{prob}{vector of probabilities}

\item{alpha}{vector of thresholds}

\item{link}{link function (logit or probit)}
}
\value{
a list of plots
}
\description{
Show the thresholds and associated probabilities
}
