% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_ord_latent}
\alias{sim_ord_latent}
\title{Simulating Ordinal Data}
\usage{
sim_ord_latent(
  location,
  scale = NULL,
  beta,
  zeta = NULL,
  prob0 = NULL,
  alpha = NULL,
  data,
  link = "logit",
  simulate = TRUE
)
}
\arguments{
\item{location}{a formula with predictors for the location parameter. The components of the formula need to be columns of the \code{data} object.}

\item{scale}{a formula with predictors for the scale parameter. Default to \code{NULL} thus if not specified the scale is assumed to be 1.}

\item{beta}{vector of regression coefficients for the location formula. The order is the same as returned by running \code{model.matrix(~formula, data = data)}}

\item{zeta}{vector of regression coefficients for the scale formula. Default to \code{NULL} thus if not specified the scale is assumed to be 1.}

\item{prob0}{vector of baseline probabilities from which thresholds are calculated. They have a role similar to intercepts in standard regression modelling. The probabilities are fixed assuming all predictors to 0.}

\item{alpha}{vector of baseline thresholds. Either \code{probs} or \code{alpha} must be specified. The function calculated thresholds from \code{probs} and viceversa}

\item{data}{a dataframe with predictors. both \code{location} and \code{scale} formulas refers to columns of the dataframe.}

\item{link}{character vector indicating the link function. Can be \code{logit} or \code{probit}.}

\item{simulate}{logical indicating if sampling values (\code{TRUE}) or returning the true probabilities}
}
\value{
a dataframe with the ordinal variable \code{y} and the latent variable \code{ys}
}
\description{
Simulating Ordinal Data
}
\examples{
# simulate a group difference in probit scale of 0.5
set.seed(123)
n <- 100
k <- 5 # 5 ordinal values
probs0 <- rep(1/k, k) # baseline probabilities for the group 0
dat <- data.frame(x = rep(c(0, 1), each = n))
dat <- sim_ord_latent(~x, beta = 0.5, prob0 = probs0, data = dat, link = "probit")
}
