% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{th_names}
\alias{th_names}
\title{Assing names to thresholds}
\usage{
th_names(k)
}
\arguments{
\item{k}{number of ordinal outcomes}
}
\description{
Give intuitive names to thresholds in the form "k - 1|k".
}
\examples{
th_names(5)
}
