% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vlogit}
\alias{vlogit}
\title{Calculate the variance of a logistic distribution}
\usage{
vlogit(scale = 1)
}
\arguments{
\item{scale}{the scale of the logistic distribution. Default to \code{1}}
}
\value{
variance of the logistic distribution
}
\description{
Calculate the variance of a logistic distribution
}
\examples{
vlogit(scale = 1)
vlogit(scale = 2)
}
